requirements-dev.txt
-r backend/requirements.plugins.txt
-r backend/requirements.txt
ipython==9.12.0
pre-commit==4.5.1
pylint-django==2.7.0
pylint==4.0.5
pytest-django==4.12.0
pytest==9.0.2
python-dotenv==1.2.2
requirementscheck==0.1.0
types-requests==2.33.0.20260327
